Remove old scaffolding code.
The idea of "index" is gone for now, though I have kept the node
holder for now.  The reasoning for keeping that is if we want to have
branch lengths, that's where they need to go, and having a wrapper
there is probably more natural than requiring all data types to have
a "length" slot, or something.