Duplicate the topology of a tree.
At the R level, this only appears to clear out the data field, which
is not very interesting.  However, at the C++ level it can build a
tree of a different type of data but the same topology (e.g.,
converting a tree<node<Robject>> -> tree<node<int>>.  Will come in useful
in a number of places.