Work towards more generic node.
I'm defining a new class node_wrapped contains a node (in the same
way that tree_wrapped contains a tree).  The idea is that if we
wrap a node with a data slot of Rcpp::RObject then provided that
the node<T> has a wrappable type T we can convert to
node_wrapped<Rcpp::RObject>.

It'd be nice to do this without going through a wrapper, but if
we target node<Rcpp::RObject> in this way then we get a recursive
set of calls to wrap (probably get-aroundable with some template
magic, but this current approach involves the least amount of
depth of understanding of how Rcpp's as/wrap and type traits work).

This will probably slow down the C <-> R translations a bit, but
saves writing a ton of code.  This will be important when we start
making trees of specific types because we should immediately be able
to start accessing them.