Progress towards a generic tree export, and exposing less of the tree
Moving away from the iterators now that this is mostly working.
Major surgery still to come.

This is mostly to support trees with exotic data types without
writing lots of boilerplate (required by the work on models) and
to let the trees be useful outside of forest and in packages that
would use it at the C++ level (like diversitree2 will).