Remove node_wrapper code
This basically undoes a7114005ff8dcf07eb7ca3834eb8cc14cf6c51e4

The approach I'd taken there does not look like it will naturally
work with Rcpp's as/wrap (which won't magically work recursively)
so it is not sufficient to have the ability to convert nodes to
automatically convert trees.  So the conversions will be more
explicit.